db updated

+++ models/Show.cs
@@ -1,11 +1,11 @@
 namespace ShowTickets.Ticketmodels{
     public class Show
     {
-        public int ShowId { get; set; }
+        public long ShowId { get; set; }
         public string Name { get; set; }
         public DateTime Date { get; set; }
 
-        public int VenueId { get; set; }
+        public long VenueId { get; set; }
         public Venue Venue { get; set; }
 
         public ICollection<ShowSeat> ShowSeats { get; set; }
